Cherry-pick phasing bug fix from fiqhack.
 - You can now phase into walls properly (before, you needed to use
   moveonly even if polymorphed into a xorn)

Not Applied:  in FIQhack, there are ways to have the phasing property
without being polymorphed into a form that can do it naturally, and
the fix fixed a bug with that, as well.  That part isn't applicable to
mainline NH4 at this time, because it relies on the properties rework
that FIQ did, which we haven't merged.

Cherry-picked from part of commit 1b5ba599b0360779675b87af0a1b30587d9c5248.