Fix a rare diff corruption issue
The sequence copy, edit (not combined), seek, edit was throwing off
the value of the coord anchor during encoding, meaning that there
was a risk of misencoded coord commands in the future (with a 1
in 85 chance with each such command). I also added some debug code
to help diagnose this sort of thing, now and in the future.