Don't crash if an undirected spell has no target
I'm reasonably sure this is one of the bugs that we have on our
list of bugs blocking 4.3-beta3. If so, great!

(In case you haven't guessed, I'm running the testsuite at the
moment; it's the easiest way to catch quick thinkos that are at
least moderately easy to trigger. Having it produce detailed
reproduction instructions and a coredump was pretty helpful.)